+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added developer troubleshooting instructions to README.
 - Add `useFullResolutionImage` to `Spatial` to allow for loading only full resolution image from pyramid.
 - Implemented ability to select an area on the Expression Histogram. On select, a new obs set selection is created. The new selection contains the ids of all obs that belong to the selected bars.
+- Add integration test for consumer site built with NextJS.
 - Implemented ability to show two columns in the feature-list view when each feature has a second identifer associated.
 - Add `CITATION.cff`
 - Added a button to recenter and rescale data to default for Scatterplot and Spatial views.